Slight cleanup - makes the code easier to read (i think). Added qunit css.
