+++ templates/qunit/qunit.html
@@ -3,7 +3,7 @@
 <head>
 <meta charset="utf-8">
 <title>QUnit Test Results for: {{filename}}</title>
-<link rel="stylesheet" href="http://code.jquery.com/qunit/qunit-1.11.0.css">
+<link rel="stylesheet" href="qunit.css">
 {{{includes}}}
 </head>
 <body>
